+++ lua/NS2Plus/CHUD_Client.lua
@@ -15,7 +15,6 @@ Script.Load("lua/NS2Plus/Client/CHUD_WeaponTime.lua")
 Script.Load("lua/NS2Plus/Client/CHUD_ViewModel.lua")
 Script.Load("lua/NS2Plus/Client/CHUD_GoldenMode.lua")
 Script.Load("lua/NS2Plus/Client/CHUD_MinimapMoveMixin.lua")
-Script.Load("lua/NS2Plus/Client/CHUD_GorgeSpit.lua")
 
 trollModes = {}