+++ lua/NS2Plus/Client/CHUD_Options.lua
@@ -1564,20 +1564,6 @@ CHUDOptions =
 				valueType = "bool",
 				sort = "A09",
 			},
-			gorgespit = {
-				name = "CHUD_GorgeSpit",
-				label = "Gorge Spit",
-				tooltip = "Replaces gorge spit with a model (rifle_grenade) so it gets outlined and is easier to see with custom alien vision.",
-				type = "select",
-				values  = { "Default", "High Visibility" },
-				defaultValue = false,
-				category = "graphics",
-				valueType = "bool",
-				sort = "A10",
-			},
-
-
-
 			sbcenter = {
 				name = "CHUD_SBCenter",
 				label = "Auto-center scoreboard",